assessment-platform/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai==0.8.6
